Refactor ScalerContext to accept a canvas provider for platform adaptation.
